QoL  for simple language and FFMPEG
Language: in UI simple mode, chosing "French" chooses the "french" prompts

FFMPEG: is we asked to create an audio file with a file name that already existed, FFMEG would just hang and our app as well.  Correction: now we increment a suffixe in the filename for the audio if it exist to fin the first one available.